requirements.txt
asgiref==3.6.0
beautifulsoup4==4.11.1
certifi==2022.12.7
charset-normalizer==3.0.1
colorama==0.4.5
Django==4.1.5
django-cors-headers==3.13.0
django-crispy-forms==1.14.0
django-filter==22.1
django-rest-framework==0.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
gunicorn==20.1.0
idna==3.4
Pillow==9.4.0
psycopg2==2.9.5
PyJWT==2.6.0
python-decouple==3.7
pytz==2022.7.1
PyYAML==6.0
qrcode==7.3.1
requests==2.28.2
soupsieve==2.3.2.post1
sqlparse==0.4.3
stripe==5.1.0
tzdata==2022.7
uritemplate==4.1.1
urllib3==1.26.14